Add the pruned.study entry related to the SNV dataset in the Profile GDS file

This function adds the names of the SNVs into the node called "pruned.study" in GDS Sample file. If a "pruned.study" entry is already present, the entry is deleted and a new entry is created.

addGDSStudyPruning(gdsProfile, pruned)

Arguments

gdsProfile

an object of class gds.class (a GDS file), the opened Profile GDS file.

pruned

a vector of character string representing the name of the SNVs.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(gdsfmt)

## Create a temporary GDS file in an test directory
gdsFilePath <- file.path(getwd(), "GDS_TEMP_1.gds")

## Only run if directory is in writing mode
if (file.access(getwd()) == 0 && !dir.exists(gdsFilePath)) {

    ## Create and open the GDS file
    tmpGDS  <- createfn.gds(filename=gdsFilePath)

    ## Vector of low allelic fraction
    study <- c("s19222", 's19588', 's19988', 's20588', 's23598')

    ## Add segments to the GDS file
    RAIDS:::addGDSStudyPruning(gdsProfile=tmpGDS, pruned=study)

    ## Read lap information from GDS file
    read.gdsn(index.gdsn(node=tmpGDS, path="pruned.study"))

    ## Close GDS file
    closefn.gds(gdsfile=tmpGDS)

    ## Delete the temporary GDS file
    unlink(x=gdsFilePath, force=TRUE)

}