Add information related to low allelic fraction associated to the SNV dataset for a specific sample into a GDS file

The function adds the information related to low allelic fraction associated to the SNV dataset for a specific sample into a GDS file, more specifically, in the "lap" node. The "lap" node must already be present in the GDS file.

addUpdateLap(gdsProfile, snpLap)

Arguments

gdsProfile

an object of class gds.class (a GDS file), a GDS file.

snpLap

a vector of numeric value representing the low allelic fraction for each SNV present in the SNV dataset. The values should be between 0 and 0.50. The length of the vector should correspond to the number of SNVs present in the "snp.id" entry of the GDS sample file.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(gdsfmt)

## Create a temporary GDS file
gdsFilePath <- file.path(tempdir(), "GDS_TEMP.gds")

## Create and open the GDS file
gdsFile  <- createfn.gds(filename=gdsFilePath)

## Create a "lap" node
add.gdsn(node=gdsFile, name="lap", val=rep(10L, 12))
sync.gds(gdsFile)

## Vector of low allelic fraction
lap <- c(0.1, 0.23, 0.34, 0.00, 0.12, 0.11, 0.33, 0.55)

## Add segments to the GDS file
RAIDS:::addUpdateLap(gdsProfile=gdsFile, snpLap=lap)
#> [1] 0

## Read lap information from GDS file
read.gdsn(index.gdsn(node=gdsFile, path="lap"))
#> [1] 0 0 0 0 0 0 0 0

## Close GDS file
closefn.gds(gdsfile=gdsFile)

## Delete the temporary GDS file
unlink(x=gdsFilePath, force=TRUE)