Append sample names into a GDS file

This function append the sample identifiers into the "samples.id" node of a GDS file.

appendGDSSampleOnly(gds, listSamples)

Arguments

gds

an object of class gds.class (a GDS file), the opened GDS file.

listSample

a vector of character string representing the sample identifiers to be added to GDS file.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Create a temporary GDS file in an test directory
dataDir <- system.file("extdata/tests", package="RAIDS")
gdsFilePath <- file.path(dataDir, "GDS_TEMP_04.gds")

## Create and open the GDS file
GDS_file_tmp  <- createfn.gds(filename=gdsFilePath)

## Create "sample.id" node (the node must be present)
add.gdsn(node=GDS_file_tmp, name="sample.id", val=c("sample_01",
    "sample_02"))

sync.gds(gdsfile=GDS_file_tmp)

## Add information about 2 samples to the GDS file
RAIDS:::appendGDSSampleOnly(gds=GDS_file_tmp,
    listSamples=c("sample_03", "sample_04"))
#> [1] 0

## Read sample identifier list
## Only "sample_03" and "sample_04" should have been added
read.gdsn(index.gdsn(node=GDS_file_tmp, path="sample.id"))
#> [1] "sample_01" "sample_02" "sample_03" "sample_04"

## Close GDS file
closefn.gds(gdsfile=GDS_file_tmp)

## Delete the temporary GDS file
unlink(x=gdsFilePath, force=TRUE)