Append information related to profile genotypes into a Population Reference GDS file (associated node already present in the GDS)

This function appends the genotype fields with the associated information into the Population Reference GDS file for the selected profiles. The associated node must already present in the GDS file.

appendGDSgenotype(gds, listSample, pathGeno, fileSNPsRDS, verbose)

Arguments

gds

an object of class gds.class (a GDS file), the opened Population Reference GDS file.

listSample

a character string representing the path and file name of the RDS file that contains the indexes of the retained SNPs. The file must exist. The file must be a RDS file.

pathGeno

a character string representing the path where the reference genotyping files for each sample are located. The name of the genotyping files must correspond to the individual identification (Individual.ID) in the pedigree file.

fileSNPsRDS

a character string representing the path and file name of the RDS file that contains the indexes of the retained SNPs. The file must exist. The file must be a RDS file.

verbose

a logical indicating if the function must print messages when running.

Value

The integer 0 when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(gdsfmt)

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## Path where the demo genotype CSV files are located
pathGeno <- file.path(dataDir, "demoProfileGenotypes")

## The RDS file containing the pedigree information
pedigreeFile <- file.path(dataDir, "PedigreeReferenceDemo.rds")

## The RDS file containing the indexes of the retained SNPs
snpIndexFile <- file.path(dataDir, "listSNPIndexes_Demo.rds")

## The RDS file containing the filtered SNP information
filterSNVFile <- file.path(dataDir, "mapSNVSelected_Demo.rds")

## Temporary Reference GDS file
tempRefGDS <- file.path(tempdir(), "Ref_TEMP02.gds")

## Create temporary Reference GDS file
newGDS <- createfn.gds(tempRefGDS)
put.attr.gdsn(newGDS$root, "FileFormat", "SNP_ARRAY")

## Read the pedigree file
ped1KG <- readRDS(pedigreeFile)

## Add information about samples to the Reference GDS file
listSampleGDS <- RAIDS:::generateGDSRefSample(gdsReference=newGDS,
                dfPedReference=ped1KG, listSamples=NULL)

## Add SNV information to the Reference GDS
RAIDS:::generateGDSSNPinfo(gdsReference=newGDS, fileFreq=filterSNVFile,
                verbose=FALSE)
#> [1] 0

## Add genotype information to the Reference GDS for the 3 first samples
RAIDS:::generateGDSgenotype(gds=newGDS, pathGeno=pathGeno,
        fileSNPsRDS=snpIndexFile, listSamples=listSampleGDS[1:3],
        verbose=FALSE)
#> [1] 0

## Append genotype information to the Reference GDS for the other samples
RAIDS:::appendGDSgenotype(gds=newGDS, pathGeno=pathGeno,
        fileSNPsRDS=snpIndexFile,
        listSample=listSampleGDS[4:length(listSampleGDS)],
        verbose=FALSE)
#> [1] 0

## Close file
closefn.gds(newGDS)

## Remove temporary files
unlink(tempRefGDS, force=TRUE)