Appends the genotype information for sepcific samples (1 column == 1 sample) into a GDS file

This function appends the genotype information into a GDS file. More specifically, the genotype information is added to the "genotype" node. The "genotype" node must already be present in the GDS file. The genotype information is a matrix with the rows corresponding to SNVs and columns corresponding to samples. The number of rows of the new genotype information must correspond to the number of rows of the matrix present in the "genotype" node.

appendGDSgenotypeMat(gds, matG)

Arguments

gds

an object of class gds.class (a GDS file), the opened GDS Sample file.

matG

a matrix of integer representing the genotypes of the SNVs for one or multiple samples. The rows correspond to SNVs and the columns correspond to samples. The number of rows must correspond to the number of rows of the matrix present in the "genotype" node.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Create a temporary GDS file in an test directory
data.dir <- system.file("extdata/tests", package="RAIDS")
gdsFilePath <- file.path(data.dir, "GDS_TEMP_06.gds")

## Create and open the GDS file
GDS_file_tmp  <- createfn.gds(filename=gdsFilePath)

## Create a "genotype" node with initial matrix
geno_initial <- matrix(rep(0L, 10), nrow=2)

add.gdsn(node=GDS_file_tmp, name="genotype", val=geno_initial)
sync.gds(GDS_file_tmp)

## New genotype information to be added
new_genotype <- matrix(rep(1L, 6), nrow=2)

## Add segments to the GDS file
RAIDS:::appendGDSgenotypeMat(gds=GDS_file_tmp, matG=new_genotype)
#> [1] 0

## Read genotype information from GDS file
## The return matrix should be a combinaison of both initial matrix
## and new matrix (column binded)
read.gdsn(index.gdsn(node=GDS_file_tmp, path="genotype"))
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]    0    0    0    0    0    1    1    1
#> [2,]    0    0    0    0    0    1    1    1

## Close GDS file
closefn.gds(gdsfile=GDS_file_tmp)

## Delete the temporary GDS file
unlink(x=gdsFilePath, force=TRUE)