TODO

TODO

computeAlleleFraction(snpPos, w = 10, cutOff = -3)

Arguments

snpPos

a data.frame containing the genotype information for a SNV dataset. TODO

w

a single positive numeric representing the size of the window to compute the allelic fraction. Default: 10.

cutOff

a single numeric representing TODO. Default: -3.

Value

a matrix of numeric with 3 columns where each row represent a segment of imbalanced SNVs. The first column represents the position, in snpPos, of the first SNV in the segment. The second column represents the position, in the snpPos, of the last SNV in the segment. The third column represents the lower allelic frequency of the segment and is NA when the value cannot be calculated. The value NULL is returned when none of the SNVs tested positive for the imbalance.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Data frame with SNV information for the specified chromosome (chr 1)
snpInfo <- data.frame(cnt.tot=c(41, 17, 27, 15, 11, 37, 16, 32),
    cnt.ref=c(40, 17, 27, 15, 4, 14, 16, 32),
    cnt.alt=c(0, 0, 0, 0, 7, 23, 0, 0),
    snp.pos=c(3722256, 3722328, 3767522, 3868160, 3869467, 4712655,
        6085318, 6213145),
    snp.chr=c(rep(1, 8)),
    normal.geno=c(rep(1, 8)),
    pruned=c(TRUE, TRUE, FALSE, TRUE, FALSE, rep(TRUE, 3)),
    snp.index=c(160, 162, 204, 256, 259, 288, 366, 465),
    keep=rep(TRUE, 8), hetero=c(rep(FALSE, 4), TRUE, TRUE, rep(FALSE, 2)),
    homo=c(rep(TRUE, 4), FALSE, FALSE, TRUE, TRUE),
    lap=rep(-1, 8), LOH=rep(0, 8), imbAR=rep(-1, 8),
    stringAsFactor=FALSE)

## The function returns NULL when there is not imbalanced SNVs
RAIDS:::computeAlleleFraction(snpPos=snpInfo, w=10, cutOff=-3)
#> NULL