Estimate the allelic fraction of the pruned SNVs for a specific DNA-seq profile

The function creates a data.frame containing the allelic fraction for the pruned SNV dataset specific to a DNA-seq profile

computeAllelicFractionDNA(
  gdsReference,
  gdsSample,
  currentProfile,
  studyID,
  chrInfo,
  minCov = 10L,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffHomoScore = -3,
  wAR = 9L,
  verbose
)

Arguments

gdsReference

an object of class gds.class (a GDS file), the opened Reference GDS file.

gdsSample

an object of class gds.class (a GDS file), the opened Profile GDS file.

currentProfile

a character string corresponding to the sample identifier as used in pruningSample function.

studyID

a character string corresponding to the name of the study as used in pruningSample function.

chrInfo

a vector of integer values representing the length of the chromosomes.

minCov

a single positive integer representing the minimum required coverage. Default: 10L.

minProb

a single numeric between 0 and 1 representing the probability that the calculated genotype call is correct. Default: 0.999.

eProb

a single numeric between 0 and 1 representing the probability of sequencing error. Default: 0.001.

cutOffLOH

a single numeric representing the cutoff, in log, for the homozygote score to assign a region as LOH. Default: -5.

cutOffHomoScore

a single numeric representing the cutoff, in log, that the SNVs in a block are called homozygote by error. Default: -3.

wAR

a single positive integer representing the size-1 of the window used to compute an empty box. Default: 9L.

verbose

a logicial indicating if the function should print message when running.

Value

a data.frame containing the allelic information for the pruned SNV dataset with coverage > minCov. The data.frame contains those columns:

cnt.tot

a integer representing the total allele count

cnt.ref

a integer representing the reference allele count

cnt.alt

a integer representing the alternative allele count

snp.pos

a integer representing the position on the chromosome

snp.chr

a integer representing the chromosome

normal.geno

a integer representing the genotype (0=wild-type reference; 1=heterozygote; 2=homozygote alternative; 3=unkown)

pruned

a logical indicating if the SNV is retained after pruning

snp.index

a integer representing the index position of the SNV in the Reference GDS file that contains all SNVs

keep

a logical indicating if the genotype exists for the SNV

hetero

a logical indicating if the SNV is heterozygote

homo

a logical indicating if the SNV is homozygote

lap

a numeric representing the lower allelic fraction

LOH

a integer indicating if the SNV is in an LOH region (0=not LOH, 1=in LOH)

imbAR

a integer indicating if the SNV is in an imbalanced region (-1=not classified as imbalanced or LOH, 0=in LOH; 1=tested positive for imbalance in at least 1 window)

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library for GDS
library(SNPRelate)

## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata/tests", package="RAIDS")
fileGDS <- file.path(dataDir, "ex1_good_small_1KG.gds")

## Temporary Profile GDS file for one profile in temporary directory
fileProfile <- file.path(tempdir(), "ex1.gds")

## Copy the Profile GDS file demo that has been pruned and annotated
file.copy(file.path(dataDir, "ex1_demo_with_pruning_and_1KG_annot.gds"),
                 fileProfile)
#> [1] TRUE

## Open the reference GDS file (demo version)
gds1KG <- snpgdsOpen(fileGDS)

## Open Profile GDS file for one profile
profileGDS <- openfn.gds(fileProfile)

## Required library for this example to run correctly
if (requireNamespace("GenomeInfoDb", quietly=TRUE) &&
     requireNamespace("BSgenome.Hsapiens.UCSC.hg38", quietly=TRUE)) {

    ## Chromosome length information
    ## chr23 is chrX, chr24 is chrY and chrM is 25
    chrInfo <- GenomeInfoDb::seqlengths(BSgenome.Hsapiens.UCSC.hg38::Hsapiens)[1:25]

    ## The function returns a data frame containing the allelic fraction info
    result <- RAIDS:::computeAllelicFractionDNA(gdsReference=gds1KG,
        gdsSample=profileGDS, currentProfile="ex1", studyID="MYDATA",
        chrInfo=chrInfo, minCov=10L,
        minProb=0.999, eProb=0.001, cutOffLOH=-5,
        cutOffHomoScore=-3, wAR=9L, verbose=FALSE)
    head(result)

    ## Close both GDS files (important)
    closefn.gds(profileGDS)
    closefn.gds(gds1KG)

    ## Remove Profile GDS file (created for demo purpose)
    unlink(fileProfile, force=TRUE)

}