Estimate the allelic fraction of the pruned SNVs for a specific RNA-seq sample

The function creates a data.frame containing the allelic fraction for the pruned SNV dataset specific to a RNA-seq sample. TODO

computeAllelicFractionRNA(
  gdsReference,
  gdsSample,
  gdsRefAnnot,
  currentProfile,
  studyID,
  blockID,
  chrInfo,
  minCov = 10L,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffAR = 3,
  verbose
)

Arguments

gdsReference

an object of class gds.class (a GDS file), the opened Reference GDS file.

gdsSample

an object of class gds.class (a GDS file), the opened Profile GDS file.

gdsRefAnnot

an object of class gds.class (a GDS file), the opened Reference SNV Annotation GDS file.

currentProfile

a character string corresponding to the sample identifier as used in pruningSample function.

studyID

a character string corresponding to the name of the study as used in pruningSample function.

blockID

a character string corresponding to the field gene block in the GDS gdsRefAnnot to use split by gene.

chrInfo

a vector of integer values representing the length of the chromosomes.

minCov

a single positive integer representing the minimum required coverage. Default: 10L.

minProb

a single numeric between 0 and 1 representing the probability that the calculated genotype call is correct. Default: 0.999.

eProb

a single numeric between 0 and 1 representing the probability of sequencing error. Default: 0.001.

cutOffLOH

a single log of the score to be LOH TODO. Default: -5.

cutOffAR

a single numeric representing the cutoff, in log score, to tag SNVs located in a gene has having an allelic fraction different 0.5 Default: 3.

verbose

a logicial indicating if the function should print message when running.

Value

a data.frame with lap for the pruned SNV dataset with coverage > minCov. TODO

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## TODO