Identify regions of LOH on one chromosome using homozygote SNVs

The function identifies regions of LOH on a specific chromosome using the homozygote SNVs present on the chromosome.

computeLOHBlocksDNAChr(gdsReference, chrInfo, snpPos, chr, genoN = 1e-04)

Arguments

gdsReference

an object of class SNPRelate::SNPGDSFileClass, an opened Reference GDS file.

chrInfo

a vector of integer representing the length of the chromosomes. As an example, the information ca be obtained from package 'BSgenome.Hsapiens.UCSC.hg38'.

snpPos

a data.frame containing the SNV information for the chromosome specified by the chr argument. The data.frame must contain:

cnt.tot a single integer representing the total coverage for the SNV.

cnt.ref a single integer representing the coverage for the reference allele.

cnt.alt a single integer representing the coverage for the alternative allele.

snp.pos a single integer representing the SNV position.

snp.chr a single integer representing the SNV chromosome.

normal.geno a single numeric indicating the genotype of the SNV. The possibles are: 0 (wild-type homozygote), 1 (heterozygote), 2 (altenative homozygote), 3 indicating that the normal genotype is unknown.

pruned a logical indicating if the SNV is retained after pruning

snp.index a integer representing the index position of the SNV in the Reference GDS file that contains all SNVs

keep a logical indicating if the genotype exists for the SNV

hetero a logical indicating if the SNV is heterozygote

homo a logical indicating if the SNV is homozygote

chr

a single positive integer for the current chromosome. The chrInfo parameter must contain the value for the specified chromosome.

genoN

a single numeric between 0 and 1 representing TODO. Default: 0.0001.

Value

a data.frame with the informations about LOH on a specific chromosome. The data.frame contains those columns:

chr a integer representing the current chromosome

start a integer representing the starting position on the box containing only homozygote SNVs (or not SNV). The first box starts at position 1.

end a integer representing the end position on the box containing only homozygote SNVs (or not SNV). The last box ends at the length of the chromosome.

logLHR TODO

LH1 TODO

LM1 TODO

homoScore a numeric representing LH1 - LM1

nbSNV a integer representing th number of SNVs in the box

nbPruned a integer representing th number of pruned SNVs in the box

nbNorm TODO

LOH TODO

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library for GDS
library(gdsfmt)

## Path to the demo Reference GDS file is located in this package
dataDir <- system.file("extdata/tests", package="RAIDS")
fileGDS <- file.path(dataDir, "ex1_good_small_1KG_GDS.gds")

## Open the Reference GDS file (demo version)
gds1KG <- snpgdsOpen(fileGDS)

## Chromosome length information for hg38
## chr23 is chrX, chr24 is chrY and chrM is 25
chrInfo <- c(248956422L, 242193529L, 198295559L, 190214555L,
    181538259L, 170805979L, 159345973L, 145138636L, 138394717L, 133797422L,
    135086622L, 133275309L, 114364328L, 107043718L, 101991189L, 90338345L,
    83257441L,  80373285L,  58617616L,  64444167L,  46709983L, 50818468L,
    156040895L, 57227415L,  16569L)

## Data frame with SNV information for the specified chromosome (chr 1)
snpInfo <- data.frame(cnt.tot=c(41, 17, 27, 15, 11, 37, 16, 32),
    cnt.ref=c(40, 17, 27, 15, 4, 14, 16, 32),
    cnt.alt=c(0, 0, 0, 0, 7, 23, 0, 0),
    snp.pos=c(3722256, 3722328, 3767522, 3868160, 3869467, 4712655,
        6085318, 6213145),
    snp.chr=c(rep(1, 8)),
    normal.geno=c(rep(3, 8)), pruned=c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE,
    TRUE, TRUE),
    pruned=c(TRUE, TRUE, FALSE, TRUE, FALSE, rep(TRUE, 3)),
    snp.index=c(160, 162, 204, 256, 259, 288, 366, 465),
    keep=rep(TRUE, 8), hetero=c(rep(FALSE, 4), TRUE, TRUE, rep(FALSE, 2)),
    homo=c(rep(TRUE, 4), FALSE, FALSE, TRUE, TRUE),
    stringAsFactor=FALSE)

## The function returns a data frame containing the information about the
## LOH regions in the specified chromosome
result <- RAIDS:::computeLOHBlocksDNAChr(gdsReference=gds1KG,
    chrInfo=chrInfo, snpPos=snpInfo, chr=1L, genoN=0.0001)
head(result)
#>   chr   start       end       logLHR  LH1 LM1 homoScore nbSNV nbPruned nbNorm
#> 1   1       1   3869467 1.734723e-17 -100   0      -100     4        3      0
#> 2   1 3869468   4712655 0.000000e+00    0   0         0     0        0      0
#> 3   1 4712656 248956422 0.000000e+00    0   0         0     2        1      0
#>   LOH
#> 1   0
#> 2   0
#> 3   0

## Close Reference GDS file (important)
closefn.gds(gds1KG)