Calculate Principal Component Analysis (PCA) on SNV genotype data set

The functions calculates the principal component analysis (PCA) for a list of pruned SNVs present in a Profile GDS file. The snpgdsPCA function is used to do the calculation.

computePCARefRMMulti(
  gdsProfile,
  refProfileIDs,
  listRM,
  np = 1L,
  algorithm = "exact",
  eigenCount = 32L,
  missingRate = 0.025,
  verbose
)

Arguments

gdsProfile

an object of class SNPGDSFileClass, the opened Profile GDS file.

refProfileIDs

a vector of reference 1KG profile identifiers that are present in the Profile GDS file. Those profiles minus the one present in the listRM vector will be used to run the PCA analysis.

listRM

a vector of character strings containing the identifiers for the reference samples that need to be removed for the PCA analysis.

np

a single positive integer representing the number of CPU that will be used. Default: 1L.

algorithm

a character string representing the algorithm used to calculate the PCA. The 2 choices are "exact" (traditional exact calculation) and "randomized" (fast PCA with randomized algorithm introduced in Galinsky et al. 2016). Default: "exact".

eigenCount

a single integer indicating the number of eigenvectors that will be in the output of the snpgdsPCA function; if 'eigenCount' <= 0, then all eigenvectors are returned. Default: 32L.

missingRate

a numeric value representing the threshold missing rate at with the SNVs are discarded; the SNVs are retained in the snpgdsPCA function with "<= missingRate" only; if NaN, no missing threshold. Default: 0.025.

verbose

a logical indicating if message information should be printed.

Value

a list containing 2 entries:

pruned a vector of SNV identifiers specifying selected SNVs for the PCA analysis.

pca.unrel a snpgdsPCAClass object containing the eigenvalues as generated by snpgdsPCA function.

References

Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ, Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72. doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo Profile GDS file is located in this package
dataDir <- system.file("extdata/demoKNNSynthetic", package="RAIDS")

# The name of the synthetic study
studyID <- "MYDATA.Synthetic"

## Profiles that should be removed from the PCA analysis
## Those profiles has been used to generate the synthetic data set
samplesRM <- c("HG00246", "HG00325", "HG00611", "HG01173", "HG02165",
    "HG01112", "HG01615", "HG01968", "HG02658", "HG01850", "HG02013",
    "HG02465", "HG02974", "HG03814", "HG03445", "HG03689", "HG03789",
    "NA12751", "NA19107", "NA18548", "NA19075", "NA19475", "NA19712",
    "NA19731", "NA20528", "NA20908")

## The known ancestry for the 1KG reference profiles
refKnownSuperPop <- readRDS(file.path(dataDir, "knownSuperPop1KG.RDS"))

## Open the Profile GDS file
gdsProfile <- snpgdsOpen(file.path(dataDir, "ex1.gds"))

## Compute PCA for the 1KG reference profiles excluding
## the profiles used to generate the synthetic profiles
results <- RAIDS:::computePCARefRMMulti(gdsProfile=gdsProfile,
    refProfileIDs=names(refKnownSuperPop), listRM=samplesRM, np=1L,
    algorithm="exact", eigenCount=32L, missingRate=0.025, verbose=FALSE)

## The PCA on the pruned SNVs data set for selected profiles
head(results$pca.unrel$eigenvect)
#>            [,1]         [,2]          [,3]        [,4]          [,5]
#> [1,] 0.01443979  0.006749237 -0.0261354543 0.004853824 -6.142577e-02
#> [2,] 0.02954388 -0.015455954 -0.0273139505 0.008782452 -2.944667e-03
#> [3,] 0.06239436  0.001841563 -0.0188625542 0.009084657 -6.499727e-05
#> [4,] 0.01662086  0.075795806 -0.0225914681 0.009334434  1.592556e-03
#> [5,] 0.05016640  0.002046317  0.0081370494 0.035458690 -4.225606e-03
#> [6,] 0.02073974  0.006225183 -0.0004003026 0.026010663 -2.542301e-02
#>              [,6]         [,7]         [,8]        [,9]        [,10]
#> [1,]  0.046466098  0.016935350  0.025488164 0.032859553  0.003045586
#> [2,]  0.011871703 -0.036909837  0.009719496 0.007039838 -0.020157239
#> [3,] -0.003012173 -0.038320544  0.008615371 0.015709646 -0.024393365
#> [4,] -0.020193363 -0.008846580 -0.036577640 0.045262360  0.002525894
#> [5,] -0.033233535 -0.027782802 -0.054582563 0.015030264  0.045112931
#> [6,] -0.020456060  0.008826842 -0.021284512 0.008237308  0.033578588
#>              [,11]        [,12]       [,13]       [,14]       [,15]       [,16]
#> [1,] -0.0097955051 -0.010897502  0.01177899 -0.01063065  0.02065920  0.02992141
#> [2,] -0.0005248565 -0.009595075  0.01906734 -0.05465565 -0.06681357 -0.02538893
#> [3,] -0.0395445523 -0.046289412 -0.04531505 -0.02062915 -0.01647287  0.03833764
#> [4,] -0.0154410216 -0.041459050 -0.04424394  0.03243328 -0.03526962  0.04619540
#> [5,] -0.0081666190 -0.023510364 -0.04197248 -0.01936231  0.03748042 -0.03718054
#> [6,]  0.0198796673 -0.015373744 -0.02631543  0.03734747  0.02324085  0.01530438
#>             [,17]        [,18]       [,19]       [,20]        [,21]
#> [1,] -0.037262097 -0.029690029  0.02830889 -0.03817532 -0.008601496
#> [2,] -0.040710137  0.001442048 -0.01437700 -0.01506918 -0.030433160
#> [3,] -0.008029339 -0.042696555  0.03316611  0.11134027 -0.030729986
#> [4,] -0.069775568 -0.045364124  0.04741745 -0.05647482 -0.113178764
#> [5,]  0.039942596 -0.017917871  0.01986519  0.07112082  0.036459651
#> [6,] -0.039247797 -0.008357042  0.01289309 -0.06181968 -0.009026734
#>             [,22]        [,23]        [,24]        [,25]         [,26]
#> [1,] -0.057061674 -0.012682448  0.061236801 -0.023344664  0.0131468596
#> [2,]  0.032523289  0.005392479  0.005269402 -0.031086903  0.0390227238
#> [3,]  0.013557873  0.025744713 -0.030705623  0.014890561 -0.0095562702
#> [4,] -0.082534309  0.086129496  0.009610947  0.008073060 -0.0330567680
#> [5,] -0.032771279 -0.011888994 -0.036764082 -0.017247722  0.0113865767
#> [6,] -0.009544941  0.030590073  0.052327577 -0.009251163 -0.0007222833
#>             [,27]       [,28]        [,29]         [,30]       [,31]
#> [1,] -0.004701625 -0.07855426 -0.014792864  0.0089382223  0.07297641
#> [2,] -0.000350641  0.02177336  0.001226259  0.0231748348  0.09829863
#> [3,]  0.011889437  0.02071010  0.018765385 -0.0001828003 -0.06999114
#> [4,]  0.044320333 -0.07387902 -0.045700943  0.2879280202 -0.18027554
#> [5,]  0.003718711  0.01959270  0.081474794  0.0162191483  0.00531908
#> [6,]  0.006273316 -0.03080420 -0.049295001  0.0292386255  0.06161742
#>             [,32]
#> [1,]  0.010894961
#> [2,] -0.009523196
#> [3,] -0.109286769
#> [4,] -0.035014731
#> [5,] -0.006395655
#> [6,] -0.059377471

## Close Profile GDS file (important)
closefn.gds(gdsProfile)