Deprecated Function

Deprecated

computePCAsynthetic(
  gdsSample,
  pruned,
  sample.id,
  sample.ref,
  study.annot,
  algorithm = "exact",
  eigen.cnt = 32L
)

Arguments

gdsSample

an object of class gds opened related to the sample

pruned

TODO

sample.id

TODO

sample.ref

TODO

study.annot

a data.frame with one entry from study.annot in the gds

algorithm

a character string representing the algorithm used to calculate the PCA. The 2 choices are "exact" (traditional exact calculation) and "randomized" (fast PCA with randomized algorithm introduced in Galinsky et al. 2016). Default: "exact".

eigen.cnt

a single integer indicating the number of eigenvectors that will be in the output of the snpgdsPCA function; if 'eigen.cnt' <= 0, then all eigenvectors are returned. Default: 32L.

Value

A list TODO with the sample.id and eigenvectors.

References

Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ, Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72. doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz