Filtering the read counts for a specific nucleotide

The function returns the read counts for the specific nucleotide or zero when read counts are not available.

extractNucleotide(nucleotide, count, curNucleo)

Arguments

nucleotide

a vector of a character strings representing the nucleotides (ex: A, C, G or T).

count

a vector of numeric representing the counts for each nucleotide listed in nucleotide parameter.

curNucleo

a character strings representing the nucleotide that will be retained (ex: A, C, G or T).

Value

a numeric representing the counts for the selected nucleotide. The default value is 0.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Nucleotides vector
nuc <- c("A", "G", "C", "T")

## Count vector
cnt <- c(100, 200, 4, 32)

## Return the count for the nucleotide "G"
RAIDS:::extractNucleotide(nucleotide=nuc, count=cnt, curNucleo="G")
#> [1] 200