TODO This function append the genotype and the file related to the pileup

TODO

generateGDS1KGgenotypeFromSNPPileup(
  PATHGENO,
  listSamples,
  listPos,
  offset,
  minCov = 10,
  minProb = 0.999,
  seqError = 0.001,
  pedStudy,
  batch,
  studyDF,
  PATHGDSSAMPLE = NULL,
  verbose = FALSE
)

Arguments

PATHGENO

a character string representing the path to a directory with the genotype files for the samples, as generated by snp-pileup. The path must exist.

listSamples

a vector of character string corresponding to the sample identifiers that will have a GDS Sample file created. The sample identifiers must be present in the "Name.ID" column of the data.frame passed to the pedDF parameter.

listPos

a array

offset

a integer to adjust if the genome start at 0 or 1

minCov

a single positive integer representing the minimum coverage needed to keep the SNPs in the analysis. Default: 10.

minProb

a array with the sample to keep

seqError

a single positive numeric between 0 and 1 representing the sequencing error rate. Default: 0.001.

batch

a single positive integer representing the current identifier for the batch. Beware, this field is not stored anymore.

studyDF

a data.frame containing the information about the study associated to the analysed sample(s). The data.frame must have those 3 columns: "study.id", "study.desc", "study.platform". All columns must be in character strings.

PATHGDSSAMPLE

TODO a PATH to a directory where a gds specific to the samples with coverage info is keep

verbose

a logical indicating if the function must print messages when running. Default: FALSE.

pedDF

a data.frame with the information about the sample(s). Those are mandatory columns: "Name.ID", "Case.ID", "Sample.Type", "Diagnosis", "Source". All columns must be in character strings. The data.frame must contain the information for all the samples passed in the listSamples parameter.

Value

The function returns 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO
gds <- "Demo GDS TODO"