Initialization of the section related to the profile information in the GDS file

This function initializes the section related to the profile information in the GDS file. The information is extracted from the data.frame passed to the function. The nodes "sample.id" and "sample.annot" are created in the GDS file.

generateGDSRefSample(gdsReference, dfPedReference, listSamples = NULL)

Arguments

gdsReference

an object of class gds.class (a GDS file), the opened GDS file.

dfPedReference

a data.frame containing the information related to the samples. It must have those columns: "sample.id", "Name.ID", "sex", "pop.group", "superPop" and "batch". All columns, except "sex" and batch", are character strings. The "batch" and "sex" columns are integer. The unique identifier of this data.frame is the "Name.ID" column. The row names of the data.frame must correspond to the identifiers present in the "Name.ID" column.

listSamples

a vector of character string representing the identifiers of the selected profiles. If NULL, all profiles are selected. Default: NULL.

Value

a vector of character string with the identifiers of the profiles saved in the GDS file.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(gdsfmt)

## Create a temporary GDS file in an test directory
dataDir <- system.file("extdata/tests", package="RAIDS")
gdsFilePath <- file.path(dataDir, "GDS_TEMP_10.gds")

## Create and open the GDS file
tmpGDS  <- createfn.gds(filename=gdsFilePath)

## Create "sample.annot" node (the node must be present)
pedInformation <- data.frame(sample.id=c("sample_01", "sample_02"),
    Name.ID=c("sample_01", "sample_02"),
    sex=c(1,1),  # 1:Male  2: Female
    pop.group=c("ACB", "ACB"),
    superPop=c("AFR", "AFR"),
    batch=c(1, 1),
    stringsAsFactors=FALSE)

## The row names must be the sample identifiers
rownames(pedInformation) <- pedInformation$Name.ID

## Add information about 2 samples to the GDS file
RAIDS:::generateGDSRefSample(gdsReference=tmpGDS,
    dfPedReference=pedInformation, listSamples=NULL)
#> [1] "sample_01" "sample_02"

## Read sample identifier list
read.gdsn(index.gdsn(node=tmpGDS, path="sample.id"))
#> [1] "sample_01" "sample_02"

## Read sample information from GDS file
read.gdsn(index.gdsn(node=tmpGDS, path="sample.annot"))
#>   sex pop.group superPop batch
#> 1   1       ACB      AFR     1
#> 2   1       ACB      AFR     1

## Close GDS file
closefn.gds(gdsfile=tmpGDS)

## Delete the temporary GDS file
unlink(x=gdsFilePath, force=TRUE)