Generate two indexes based on gene annotation for gdsAnnot1KG block

Generate two indexes based on gene annotation for gdsAnnot1KG block

generateGeneBlock(gdsReference, winSize = 10000, ensDb)

Arguments

gdsReference

an object of class gds.class (a GDS file), the opened 1KG GDS file (reference).

winSize

a single positive integer representing the size of the window to use to group the SNVs when the SNVs are in a non-coding region. Default: 10000.

ensDb

An object of class EnsDb with the Ensembl genome annotation. By default, the EnsDb.Hsapiens.v86 class has been used.

Value

a data.frame with those columns:

chr

a single integer representing the SNV chromosome.

pos

a single integer representing the SNV position.

snp.allele

a character string representing the reference allele and alternative allele for each of the SNV

Exon

a character with the ensembl GeneId(s) if the SNV is in one exon. If more than one GeneId they are separted by ':'

GName

a character with the ensembl GeneId(s) if the SNV is in the gene. If more than one GeneId they are separted by ':'

Gene

a single integer specific to the SNVs that share at least one genes

GeneS

a single integer specific to the SNVs that share a unique combination of genes

"chr", "pos", "snp.allele", "Exon", "GName", "Gene", "GeneS" Example for GName and the two indexes "Gene", "GeneS" GName Gene GeneS 470 ENSG00000230021 17 3820 471 ENSG00000230021 17 3820 472 ENSG00000230021:ENSG00000228794 17 3825 473 ENSG00000230021:ENSG00000228794 17 3825 481 ENSG00000230021:ENSG00000228794:ENSG00000225880 17 3826 482 ENSG00000230021:ENSG00000228794:ENSG00000225880 17 3826 483 ENSG00000230021:ENSG00000228794:ENSG00000225880 17 3826 492 ENSG00000230021:ENSG00000228794 17 3825 493 ENSG00000230021:ENSG00000228794 17 3825

Author

Pascal Belleau, Astrid Deschênes and Alex Krasnitz

Examples

## Required library
library(SNPRelate)

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## Required library
if (requireNamespace("EnsDb.Hsapiens.v86", quietly=TRUE)) {

    ## Making a "short cut" on the ensDb object
    edb <- EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86

    path1KG <- file.path(dataDir, "tests")

    ## Reference GDS file
    fileReferenceGDS  <- file.path(path1KG, "ex1_good_small_1KG.gds")

    # \donttest{
        ## Open the reference GDS file (demo version)
        gds1KG <- snpgdsOpen(fileReferenceGDS)

        ## The function returns a data.frame containing
        ## gene block information
        matGeneBlock <- RAIDS:::generateGeneBlock(gdsReference=gds1KG,
                            ensDb=edb)
        print(head(matGeneBlock[grep("ENSG00000157152",
                            matGeneBlock$GName),]))

        closefn.gds(gds1KG)
    # }
}
#>      chr      pos snp.allele            Exon                           GName
#> 1311   3 12070459        C/T ENSG00000226642 ENSG00000157152:ENSG00000226642
#> 1312   3 12070683        G/A ENSG00000226642 ENSG00000157152:ENSG00000226642
#> 1313   3 12070758        G/A ENSG00000226642 ENSG00000157152:ENSG00000226642
#> 1314   3 12145519        A/G                                 ENSG00000157152
#> 1315   3 12146031        G/A                                 ENSG00000157152
#> 1316   3 12154386        G/A ENSG00000157150 ENSG00000157152:ENSG00000157150
#>      Gene GeneS
#> 1311  406    83
#> 1312  406    83
#> 1313  406    83
#> 1314  406    84
#> 1315  406    84
#> 1316  406    82