Extract the genotype information for a SNV dataset using the Profile GDS file and the Reference GDS file

The function generates a data.frame containing the genotype information from a initial list of SNVs associated to a specific profile. The function uses the information present in the Reference GDS file and the Profile GDS file.

getTableSNV(
  gdsReference,
  gdsSample,
  currentProfile,
  studyID,
  minCov = 10,
  minProb = 0.999,
  eProb = 0.001,
  verbose
)

Arguments

gdsReference

an object of class gds.class (a GDS file), the opened Reference GDS file.

gdsSample

an object of class gds.class (a GDS file), the opened Profile GDS file.

currentProfile

a character string corresponding to the sample identifier used in pruningSample function.

studyID

a character string corresponding to the study identifier used in pruningSample function.

minCov

a single positive integer representing the minimum required coverage. Default: 10L.

minProb

a single numeric between 0 and 1 representing the probability that the calculated genotype call is correct. Default: 0.999.

eProb

a single numeric between 0 and 1 representing the probability of sequencing error. Default: 0.001.

verbose

a logicial indicating if messages should be printed when the function is running.

Value

a data.frame containing:

cnt.tot

a single integer representing the total coverage for the SNV.

cnt.ref

a single integer representing the coverage for the reference allele.

cnt.alt

a single integer representing the coverage for the alternative allele.

snpPos

a single integer representing the SNV position.

snp.chr

a single integer representing the SNV chromosome.

normal.geno

a single numeric indicating the genotype of the SNV. The possibles are: 0 (wild-type homozygote), 1 (heterozygote), 2 (altenative homozygote), 3 indicating that the normal genotype is unknown.

pruned

a logical

snp.index

a vector of integer representing the position of the SNVs in the Reference GDS file.

keep

a logical

hetero

a logical

homo

a logical

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(gdsfmt)

## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata/tests", package="RAIDS")
fileGDS <- file.path(dataDir, "ex1_good_small_1KG.gds")

## Temporary Profile GDS file for one profile in temporary directory
fileProfile <- file.path(tempdir(), "ex1.gds")

## Copy the Profile GDS file demo that has been pruned and annotated
file.copy(file.path(dataDir, "ex1_demo_with_pruning_and_1KG_annot.gds"),
                 fileProfile)
#> [1] TRUE

## Open the reference GDS file (demo version)
gds1KG <- snpgdsOpen(fileGDS)

## Open Profile GDS file for one profile
profileGDS <- openfn.gds(fileProfile)

## The function returns a data frame containing the SNVs information
result <- RAIDS:::getTableSNV(gdsReference=gds1KG, gdsSample=profileGDS,
        currentProfile="ex1", studyID="MYDATA", minCov=10L, minProb=0.999,
        eProb=0.001, verbose=FALSE)
head(result)
#>   cnt.tot cnt.ref cnt.alt snp.pos snp.chr normal.geno pruned snp.index keep
#> 1      41      40       0 3722256       1           3   TRUE       160 TRUE
#> 2      17      17       0 3722328       1           3   TRUE       162 TRUE
#> 3      27      27       0 3767522       1           3  FALSE       204 TRUE
#> 4      15      15       0 3868160       1           3   TRUE       256 TRUE
#> 5      11       4       7 3869467       1           3  FALSE       259 TRUE
#> 6      37      14      23 4712655       1           3   TRUE       288 TRUE
#>   hetero  homo
#> 1  FALSE  TRUE
#> 2  FALSE  TRUE
#> 3  FALSE  TRUE
#> 4  FALSE  TRUE
#> 5   TRUE FALSE
#> 6   TRUE FALSE

## Close both GDS files (important)
closefn.gds(profileGDS)
closefn.gds(gds1KG)

## Remove Profile GDS file (created for demo purpose)
unlink(fileProfile, force=TRUE)