Find the pruned snv in 1KG by chr

TODO

pruning1KG.Chr(
  gds,
  method = "corr",
  listSamples = NULL,
  slide.max.bp.v = 5e+05,
  ld.threshold.v = sqrt(0.1),
  np = 1,
  verbose.v = FALSE,
  chr = NULL,
  minAF = NULL,
  outPref = "pruned_1KG",
  keepObj = FALSE
)

Arguments

gds

an object of class SNPRelate::SNPGDSFileClass, a SNP GDS file.

method

a character string TODO . Default: "corr".

listSamples

TODO

slide.max.bp.v

TODO

ld.threshold.v

TODO

np

TODO . Default: NULL.

verbose.v

a logical specifying if the function must provide more information about the process. Default: FALSE.

chr

TODO

minAF

TODO

outPref

TODO

keepObj

a logical specifying if the function must save the the processed information into a RDS object. Default: FALSE.

Value

NULL invisibly.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO