Compute the list of pruned SNVs for a specific sample using the information from the 1KG GDS file and the

Compute the list of pruned SNVs for a specific sample. When a group of SNVs are in linkage disequilibrium, only one SNV from that group is retained. The linkage disequilibrium is calculated with the snpgdsLDpruning() function.

pruningSample(
  gds,
  method = c("corr", "r", "dprime", "composite"),
  sampleCurrent,
  study.id,
  listSNP = NULL,
  slide.max.bp.v = 500000L,
  ld.threshold.v = sqrt(0.1),
  np = 1,
  verbose.v = FALSE,
  chr = NULL,
  minAF.SuperPop = NULL,
  keepGDSpruned = TRUE,
  PATHSAMPLEGDS = NULL,
  keepFile = FALSE,
  PATHPRUNED = ".",
  outPref = "pruned"
)

Arguments

gds

an object of class gdsn.class, a GDS node pointing to the 1KG GDS file.

method

a character string that represents the method that will be used to calculate the linkage disequilibrium in the snpgdsLDpruning() function. The 4 possible values are: "corr", "r", "dprime" and "composite". Default: "corr".

sampleCurrent

a character string corresponding to the sample identifier used in LD pruning done by the snpgdsLDpruning() function.

study.id

A string corresponding to the study use in LDpruning

listSNP

the list of snp.id keep. TODO. Default: NULL.

slide.max.bp.v

a single positive integer that represents the maximum basepairs (bp) in the sliding window. This parameter is used for the LD pruning done in the snpgdsLDpruning function. Default: 500000L.

ld.threshold.v

a single numeric value that represents the LD threshold used in the snpgdsLDpruning function. Default: sqrt(0.1).

np

a single positive integer specifying the number of threads to be used. Default: 1L.

verbose.v

a logicial indicating if information is shown during the process in the snpgdsLDpruning function. Default: FALSE.

chr

TODO. Default: NULL.

minAF.SuperPop

TODO. Default: NULL.

keepGDSpruned

a logicial TODO. Default: TRUE.

PATHSAMPLEGDS

TODO

keepFile

a logical indicating if RDS files containing the information about the pruned SNVs must be created. Default: FALSE.

PATHPRUNED

a character string TODO. Default: ".".

outPref

a character string that represents the prefix of the RDS files that will be generated. The RDS files are only generated when the parameter keepFile=TRUE. Default: "pruned".

Value

The function returns 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO