Read a SNP-pileup file

The function reads a generic SNP pileup file and returns a data frame containing the information about the read counts for the SNVs present in the file.

readSNVPileupFile(fileName, offset = 0L)

Arguments

fileName

a character string representing the name, including the path, of a text file containing the SNV read counts as generated by snp-pileup software. The text file must be comma separated. The text file must contain those columns: Chromosome, Position, Ref, Alt, File1R, File1A, File1E and File1D.

offset

a integer representing the offset to be added to the position of the SNVs. The value of offset is added to the position present in the file. Default: 0L.

Value

the a data.frame containing at least:

Chromosome

a numeric representing the name of the chromosome

Position

a numeric representing the position on the chromosome

Ref

a character string representing the reference nucleotide

Alt

a character string representing the alternative nucleotide

File1R

a numeric representing the count for the reference nucleotide

File1A

a numeric representing the count for the alternative nucleotide

File1E

a numeric representing the count for the errors

File1D

a numeric representing the count for the deletions

count

a numeric representing the total count

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Directory where demo SNP-pileup file
dataDir <- system.file("extdata/example/snpPileup", package="RAIDS")

## The SNP-pileup file
snpPileupFile <- file.path(dataDir, "ex1.txt.gz")

info <- RAIDS:::readSNVPileupFile(fileName=snpPileupFile)
head(info)
#>   Chromosome Position Ref Alt File1R File1A File1E File1D count
#> 1          1  3467333   A   G      0     37      0      0    37
#> 2          1  3467428   C   G     10      0      0      0    10
#> 3          1  3469375   C   T     13      0      0      0    13
#> 4          1  3469387   C   T     16      0      0      0    16
#> 5          1  3469502   T   G     28      0      0      0    28
#> 6          1  3469527   C   T     32      0      0      0    32