Identity-by-descent (IBD) analysis

This function calculates the IDB coefficients by KING method of moment using the SNPRelate::snpgdsIBDKING function.

runIBDKING(gds, profileID = NULL, snpID = NULL, maf = 0.05, verbose)

Arguments

gds

an object of class SNPRelate::SNPGDSFileClass, an opened SNP GDS file.

profileID

a vector of character strings representing the samples to keep for the analysis. If NULL, all samples are used. Default: NULL.

snpID

a vector of character strings representing the SNPs to keep for the analysis. If NULL, all SNPs are used. Default: NULL.

maf

a single numeric representing the threshold for the minor allele frequency. Only the SNPs with ">= maf" are retained. Default: 0.05.

verbose

a logical indicating if information is shown during the process in the snpgdsIBDKING() function.

Value

a list containing:

sample.ida character string representing the sample ids used in the analysis

snp.ida character string representing the SNP ids used in the analysis

k0a numeric, the IBD coefficient, the probability of sharing zero IBD

k1a numeric, the IBD coefficient, the probability of sharing one IBD

IBS0a numeric, the proportion of SNPs with zero IBS

kinshipa numeric, the proportion of SNPs with zero IBS, if the parameter kinship=TRUE

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required
library(SNPRelate)

## Open an example dataset (HapMap)
genoFile <- snpgdsOpen(snpgdsExampleFileName())

## Extract CEU population
samples <- read.gdsn(index.gdsn(genoFile, "sample.id"))
CEU <- samples[
    read.gdsn(index.gdsn(genoFile, "sample.annot/pop.group"))=="CEU"]

## Infer the presence of population stratification
ibd.robust <- RAIDS:::runIBDKING(gds=genoFile, profileID=CEU, snpID=NULL,
     maf=0.05, verbose=FALSE)

## close the genotype file
snpgdsClose(genoFile)