SNP pruning based on linkage disequilibrium (LD)

This function is a wrapper for the snpgdsLDpruning() function that generates a pruned subset of SNPs that are in approximate linkage equilibrium.

runLDPruning(
  gds,
  method = c("corr", "r", "dprime", "composite"),
  listSamples = NULL,
  listKeep = NULL,
  slideWindowMaxBP = 500000L,
  thresholdLD = sqrt(0.1),
  np = 1L,
  verbose.v = FALSE
)

Arguments

gds

an object of class SNPGDSFileClass, a SNP GDS file.

method

a character string that represents the method that will be used to calculate the linkage disequilibrium in the snpgdsLDpruning() function. The 4 possible values are: "corr", "r", "dprime" and "composite". Default: "corr".

listSamples

a vector of character strings corresponding to the sample identifiers used in LD pruning done by the snpgdsLDpruning() function. If NULL, all samples are used. Default: NULL.

listKeep

a vector of SNVs identifiers specifying selected; if NULL, all SNVs are used in the snpgdsLDpruning function. Default: NULL.

slideWindowMaxBP

a single positive integer that represents the maximum basepairs (bp) in the sliding window. This parameter is used for the LD pruning done in the snpgdsLDpruning() function. Default: 500000L.

thresholdLD

a single numeric value that represents the LD threshold used in the snpgdsLDpruning function. Default: sqrt(0.1).

np

a single positive integer specifying the number of threads to be used. Default: 1L.

verbose.v

a logical indicating if information is shown during the process in the snpgdsLDpruning() function. Default: FALSE.

Value

a list of SNP identifiers stratified by chromosomes as generated by snpgdsLDpruning function.

Details

The SNP pruning is based on linkage disequilibrium (LD) and is done by the snpgdsLDpruning() function in the SNPRelate package (https://bioconductor.org/packages/SNPRelate/).

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO
gds <- "Demo GDS TODO"