Compile all the inferred ancestry results done on the synthetic profiles for different D and K values in the objective of selecting the optimal D and K values for a specific profile

The function calculates the accuracy of the inferred ancestry called done on the synthetic profiles for different D and K values. The accuracy is also calculated for each super-population used to generate the synthetic profiles. The known ancestry from the reference profiles used to generate the synthetic profiles is required to calculate the accuracy.

selParaPCAUpQuartile(
  matKNN,
  pedCall,
  refCall,
  predCall,
  listCall,
  kList = seq(3, 15, 1),
  pcaList = seq(2, 15, 1)
)

Arguments

matKNN

a data.frame containing the inferred ancestry for the synthetic profiles for different K and D values. The data.frame must contained those columns: "sample.id", "D", "K" and the fourth column name must correspond to the predCall argument.

pedCall

a data.frame containing the information about the super-population information from the 1KG GDS file for profiles used to generate the synthetic profiles. The data.frame must contained a column named as the refCall argument.

refCall

a character string representing the name of the column that contains the known ancestry for the reference profiles in the Reference GDS file.

predCall

a character string representing the name of the column that contains the inferred ancestry for the specified profiles. The column must be present in the matKNN data.frame argument.

listCall

a vector of character strings representing the list of possible ancestry assignations.

kList

a vector of integer representing the list of values tested for the K parameter. The K parameter represents the number of neighbors used in the K-nearest neighbor analysis. Default: seq(3,15,1).

pcaList

a vector of integer representing the list of values tested for the D parameter. The D parameter represents the number of dimensions used in the PCA analysis. Default: seq(2,15,1).

Value

a list containing 5 entries:

dfPCA

a data.frame containing statistical results on all combined synthetic results done with a fixed value of D (the number of dimensions). The data.frame contains those columns:

D

a numeric representing the value of D (the number of dimensions).

median

a numeric representing the median of the minimum AUROC obtained (within super populations) for all combination of the fixed D value and all tested K values.

mad

a numeric representing the MAD of the minimum AUROC obtained (within super populations) for all combination of the fixed D value and all tested K values.

upQuartile

a numeric representing the upper quartile of the minimum AUROC obtained (within super populations) for all combination of the fixed D value and all tested K values.

k

a numeric representing the optimal K value (the number of neighbors) for a fixed D value.

dfPop

a data.frame containing statistical results on all combined synthetic results done with different values of D (the number of dimensions) and K (the number of neighbors). The data.frame contains those columns:

D

a numeric representing the value of D (the number of dimensions).

K

a numeric representing the value of K (the number of neighbors).

AUROC.min

a numeric representing the minimum accuracy obtained by grouping all the synthetic results by super-populations, for the specified values of D and K.

AUROC

a numeric representing the accuracy obtained by grouping all the synthetic results for the specified values of D and K.

Accu.CM

a numeric representing the value of accuracy of the confusion matrix obtained by grouping all the synthetic results for the specified values of D and K.

D

a numeric representing the optimal D value (the number of dimensions) for the specific profile.

K

a numeric representing the optimal K value (the number of neighbors) for the specific profile.

listD

a numeric representing the optimal D values (the number of dimensions) for the specific profile. More than one D is possible.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Loading demo dataset containing pedigree information for synthetic
## profiles and known ancestry of the profiles used to generate the
## synthetic profiles
data(pedSynthetic)

## Loading demo dataset containing the inferred ancestry results
## for the synthetic data
data(matKNNSynthetic)

## Compile all the results for ancestry inference done on the
## synthetic profiles for different D and K values
## Select the optimal D and K values
results <- RAIDS:::selParaPCAUpQuartile(matKNN=matKNNSynthetic,
    pedCall=pedSynthetic, refCall="superPop", predCall="SuperPop",
    listCall=c("EAS", "EUR", "AFR", "AMR", "SAS"), kList=seq(3,15,1),
    pcaList=seq(2,15,1))
results$D
#> [1] 7
results$K
#> [1] 8