A small data.frame containing the SNV information.

The object is a data.frame with 17 columns.

data(snpPositionDemo)

Format

The data.frame containing the information about the synthetic profiles. The data.frame contains 4 columns:

cnt.tot

a integer representing the number of reads at the SNV position.

cnt.ref

a integer representing the number of reads corresponding to the reference at the SNV position.

cnt.alt

a integer representing the number of reads different than the reference at the SNV position.

snp.pos

a integer representing the position of the SNV on the chromosome.

snp.chr

a integer representing the chromosome on which the SNV is located.

normal.geno

a integer representing the genotype (0=wild-type reference; 1=heterozygote; 2=homozygote alternative; 3=unkown).

pruned

a logical indicated if the SNV is pruned.

snp.index

a integer representing the index of the SNV in the reference SNV GDS file.

keep

a logical indicated if the genotype exists for the SNV.

hetero

a logical indicated if the SNV is heterozygote.

homo

a logical indicated if the SNV is homozygote.

block.id

a integer representing the block identifier associated to the current SNV.

phase

a integer representing the block identifier associated to the current SNV.

lap

a numeric representing the lower allelic fraction.

LOH

a integer indicating if the SNV is in an LOH region (0=not LOH, 1=in LOH).

imbAR

a integer indicating if the SNV is in an imbalanced region (-1=not classified as imbalanced or LOH, 0=in LOH; 1=tested positive for imbalance in at least 1 window).

freq

a numeric representing the frequency of the variant in the the reference.

Value

The data.frame containing the information about the synthetic profiles. The data.frame contains 4 columns:

cnt.tot

a integer representing the number of reads at the SNV position.

cnt.ref

a integer representing the number of reads corresponding to the reference at the SNV position.

cnt.alt

a integer representing the number of reads different than the reference at the SNV position.

snp.pos

a integer representing the position of the SNV on the chromosome.

snp.chr

a integer representing the chromosome on which the SNV is located.

normal.geno

a integer representing the genotype (0=wild-type reference; 1=heterozygote; 2=homozygote alternative; 3=unkown).

pruned

a logical indicated if the SNV is pruned.

snp.index

a integer representing the index of the SNV in the reference SNV GDS file.

keep

a logical indicated if the genotype exists for the SNV.

hetero

a logical indicated if the SNV is heterozygote.

homo

a logical indicated if the SNV is homozygote.

block.id

a integer representing the block identifier associated to the current SNV.

phase

a integer representing the block identifier associated to the current SNV.

lap

a numeric representing the lower allelic fraction.

LOH

a integer indicating if the SNV is in an LOH region (0=not LOH, 1=in LOH).

imbAR

a integer indicating if the SNV is in an imbalanced region (-1=not classified as imbalanced or LOH, 0=in LOH; 1=tested positive for imbalance in at least 1 window).

freq

a numeric representing the frequency of the variant in the the reference.

Details

This dataset can be used to test the calcAFMLRNA and tableBlockAF internal functions.

Examples

## Loading demo dataset containing SNV information
data(snpPositionDemo)

## Only use a subset of heterozygote SNVs related to one block
subset <- snpPositionDemo[which(snpPositionDemo$block.id == 2750 &
                  snpPositionDemo$hetero), c("cnt.ref", "cnt.alt", "phase")]

## Compute the log likelihood ratio based on the coverage of
## each allele in a specific block
result <- RAIDS:::calcAFMLRNA(subset)
head(result)
#> $lR
#> [1] 4.057862
#> 
#> $aFraction
#> [1] 0.3389831
#> 
#> $nPhase
#> [1] 0
#> 
#> $sumAlleleLow
#> [1] 60
#> 
#> $sumAlleleHigh
#> [1] 117
#>