Generate synthetic profiles for each cancer sample and 1KG reference sample combination and add them to the GDS Sample file

The functions uses one cancer sample in combination with one 1KG reference sample to generate an synthetic profile that is saved in the GDS Sample file.

When more than one 1KG reference samples are specified, the function recursively generates synthetic profiles for each cancer sample + 1KG reference sample combination.

The number of synthetic profiles generated by combination is specified by the number of simulation requested.

syntheticGeno(
  gds,
  gdsRefAnnot,
  gdsSampleFile,
  data.id.profile,
  listSampleRef,
  nbSim = 1L,
  prefId = "",
  pRecomb = 0.01,
  minProb = 0.999,
  seqError = 0.001
)

Arguments

gds

an object of class gds.class (a GDS file), the 1KG GDS file.

gdsRefAnnot

an object of class gds.class (a GDS file), the1 1KG SNV Annotation GDS file.

gdsSampleFile

a character string representing the file name of the GDS Sample file containing the information about the sample. The file must exist.

data.id.profile

a character string representing the unique identifier of the cancer sample.

listSampleRef

a vector of character strings representing the sample identifiers of the 1KG selected reference samples.

nbSim

a single positive integer representing the number of simulations that will be generated per sample + 1KG reference combination. Default: 1L.

prefId

a character string that represent the prefix that will be added to the name of the synthetic profiles generated by the function. Default: "".

pRecomb

a single positive numeric between 0 and 1 that represents the frequency of phase switching in the synthetic profiles, Default: 0.01.

minProb

a single positive numeric between 0 and 1 that represents the probability that the genotype is correct. TODO. Default: 0.999.

seqError

a single positive numeric between 0 and 1 representing the sequencing error rate. Default: 0.001.

Value

The integer OL when the function is successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

# TODO