Calculate presence of allelic fraction change using specified SNVs separately and together

The function tests allelic fraction change using all specified SNVs separately and together. The function reports the associated results, including statistic for the region represented by all the SNVs.

testAlleleFractionChange(matCov, pCutOff = -3, vMean)

Arguments

matCov

a data.frame containing only heterozygote SNVs. The data.frame must contain those columns:

cnt.ref a single integer representing the coverage for the reference allele.

cnt.alt a single integer representing the coverage for the alternative allele.

pCutOff

a numeric representing the cut-off for considering a region imbalanced when comparing likelihood to gave allelic fraction change and likelihood not to have allelic fraction change. Default: -3.

vMean

a positive numeric representing the current ratio (minor allele/(minor allele + second allele)) that is going to be used as reference to see if there is a allelic fraction change.

Value

a list containing 4 entries:

pWin a vector of numeric representing the probability (x2) of obtaining the current alternative/(alternative+reference) ratio from a reference distribution specified by user.

pa integer indicating if all SNVs tested positive (1=TRUE, 0=FALSE). The cut-off is 0.5.

pCuta integer indicating if all SNVs tested positive (1=TRUE, 0-FALSE).

pCut1a integer indicating if the region tested positive (1=TRUE, 0=FALSE) for allelic ratio change.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Data frame with SNV information
snpInfo <- data.frame(
    cnt.ref=c(40, 17, 27, 15, 4, 14, 16, 32),
    cnt.alt=c(2, 4, 5, 10, 7, 23, 0, 0))

RAIDS:::testAlleleFractionChange(matCov=snpInfo, pCutOff=-3, vMean=0.5)
#> $pWin
#> [1] 2.153229e-10 4.343510e-03 6.618770e-05 3.269396e-01 3.876953e-01
#> [6] 1.433067e-01 1.525879e-05 2.328306e-10
#> 
#> $p
#> [1] -11.74078
#> 
#> $pCut
#> [1] 1
#> 
#> $pCut1
#> [1] 1
#>