Calculate imbalance region using specified heterozygote SNVs separately and together

The function tests imbalance region using all specified SNVs separately and together. The function reports the associated results, including statistic for the region.

testEmptyBox(matCov, pCutOff = -3)

Arguments

matCov

a data.frame containing only heterozygote SNVs. The data.frame must contain those columns:

cnt.ref a single integer representing the coverage for the reference allele.

cnt.alt a single integer representing the coverage for the alternative allele.

pCutOff

a numeric representing the cut-off for considering a region imbalanced when comparing likelihood to be imbalanced and likelihood not to be imbalanced. Default: -3.

Value

a list containing 4 entries:

pWin a vector of numeric representing the probability (x2) of obtaining the current alternative/(alternative+reference) ratio from a 0.5 distribution.

p a numeric representing the likelihood for the region

pCut a integer indicating if all SNVs tested positive (1=TRUE, 0=FALSE). The cut-off is 0.5.

pCut1 a integer indicating if the window tested positive (1=TRUE, 0=FALSE) for imbalance.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Data frame with SNV information for a list of heterozygote SNVs
snpInfo <- data.frame(
    cnt.ref=c(40, 17, 27, 15, 4, 14, 16, 32),
    cnt.alt=c(2, 4, 5, 10, 7, 23, 0, 0))

## Calculate imbalance for the region represented by the SNVs
RAIDS:::testEmptyBox(matCov=snpInfo, pCutOff=-3)
#> $pWin
#> [1] 4.110916e-10 7.197380e-03 1.130742e-04 4.243562e-01 5.488281e-01
#> [6] 1.877416e-01 3.051758e-05 4.656613e-10
#> 
#> $p
#> [1] -12.26236
#> 
#> $pCut
#> [1] 0
#> 
#> $pCut1
#> [1] 1
#>