Validate the parameters for the addStudy1Kg() function

The function validates the input parameters for the addStudy1Kg function. When a parameter is not as expected, an error message is generated.

validateAddStudy1Kg(gdsReference, fileProfileGDS, verbose)

Arguments

gdsReference

an object of class gds.class (a GDS file), the opened 1KG GDS file.

fileProfileGDS

a character string representing the path and file name of the GDS Sample file. The GDS Sample file must exist.

verbose

a logical indicating if messages should be printed to show how the different steps in the function.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata", package="RAIDS")
fileReferenceGDS <- file.path(dataDir, "1KG_Demo.gds")
gds1KG <- snpgdsOpen(fileReferenceGDS)

## Path to demo Profile GDS file
fileProfileGDS  <- file.path(dataDir, "GDS_Sample_with_study_demo.gds")

## Returns OL when all parameters are valid
RAIDS:::validateAddStudy1Kg(gdsReference=gds1KG,
    fileProfileGDS=fileProfileGDS, verbose=TRUE)
#> [1] 0

## All GDS file must be closed
closefn.gds(gdsfile=gds1KG)