Validate the input parameters for computeKNNRefSample() function

The function validates the input parameters for the computeKNNRefSample function. When a parameter is not as expected, an error message is generated.

validateComputeKNNRefSample(
  listEigenvector,
  listCatPop,
  spRef,
  fieldPopInfAnc,
  kList,
  pcaList
)

Arguments

listEigenvector

a list with 3 entries: 'sample.id', 'eigenvector.ref' and 'eigenvector'. The list represents the PCA done on the 1KG reference profiles and the specific profile projected onto it. The 'sample.id' entry must contain only one identifier (one profile).

listCatPop

a vector of character string representing the list of possible ancestry assignations.

spRef

vector of character strings representing the known super population ancestry for the 1KG profiles. The 1KG profile identifiers are used as names for the vector.

fieldPopInfAnc

a character string representing the name of the column that will contain the inferred ancestry for the specified dataset.

kList

a vector of integer representing the list of values tested for the K parameter. The K parameter represents the number of neighbors used in the K-nearest neighbors analysis.

pcaList

a vector of integer representing the list of values tested for the D parameter. The D parameter represents the number of dimensions used in the PCA analysis.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Load the demo PCA on the synthetic profiles projected on the
## demo 1KG reference PCA
data(demoPCASyntheticProfiles)

## Load the known ancestry for the demo 1KG reference profiles
data(demoKnownSuperPop1KG)

pcaSynthetic <- demoPCASyntheticProfiles
pcaSynthetic$sample.id <- pcaSynthetic$sample.id[1]

## The function returns 0L when all parameters are valid
RAIDS:::validateComputeKNNRefSample(listEigenvector=pcaSynthetic,
    listCatPop=c("EAS", "EUR", "AFR", "AMR", "SAS"),
    spRef=demoKnownSuperPop1KG, fieldPopInfAnc="Superpop",
    kList=c(10, 11, 12), pcaList=c(13, 14, 15))
#> [1] 0