Validate the input parameters for computeKNNRefSynthetic() function

The function validates the input parameters for the computeKNNRefSynthetic function. When a parameter is not as expected, an error message is generated.

validateComputeKNNRefSynthetic(
  gdsProfile,
  listEigenvector,
  listCatPop,
  studyIDSyn,
  spRef,
  fieldPopInfAnc,
  kList,
  pcaList
)

Arguments

gdsProfile

an object of class SNPRelate::SNPGDSFileClass, the opened Profile GDS file.

listEigenvector

a list with 3 entries: 'sample.id', 'eigenvector.ref' and 'eigenvector'. The list represents the PCA done on the 1KG reference profiles and the synthetic profiles projected onto it.

listCatPop

a vector of character string representing the list of possible ancestry assignations.

studyIDSyn

a character string corresponding to the study identifier. The study identifier must be present in the Profile GDS file.

spRef

vector of character strings representing the known super population ancestry for the 1KG profiles. The 1KG profile identifiers are used as names for the vector.

fieldPopInfAnc

a character string representing the name of the column that will contain the inferred ancestry for the specified dataset.

kList

a vector of integer representing the list of values tested for the K parameter. The K parameter represents the number of neighbors used in the K-nearest neighbors analysis.

pcaList

a vector of integer representing the list of values tested for the D parameter. The D parameter represents the number of dimensions used in the PCA analysis.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo GDS file is located in this package
dataDir <- system.file("extdata/demoKNNSynthetic", package="RAIDS")
fileProfileGDS <- file.path(dataDir, "ex1.gds")

pcaSynthetic <- readRDS(file.path(dataDir, "pcaSynthetic.RDS"))

## The known ancestry for the 1KG reference profiles
refKnownSuperPop <- readRDS(file.path(dataDir, "knownSuperPop1KG.RDS"))

## Open GDS files
gdsProfile <- openfn.gds(fileProfileGDS)

## The function returns 0L when all parameters are valid
RAIDS:::validateComputeKNNRefSynthetic(gdsProfile=gdsProfile,
    listEigenvector=pcaSynthetic,
    listCatPop=c("EAS", "EUR", "AFR", "AMR", "SAS"),
    studyIDSyn="MyStudy", spRef=refKnownSuperPop,
    fieldPopInfAnc="Superpop", kList=c(10, 11, 12),
    pcaList=c(13, 14, 15))
#> [1] 0

## Close GDS file (it is important to always close the GDS files)
closefn.gds(gdsProfile)