Validate the input parameters for computePCAMultiSynthetic() function

The function validates the input parameters for the computePCAMultiSynthetic function. When a parameter is not as expected, an error message is generated.

validateComputePCAMultiSynthetic(
  gdsProfile,
  listPCA,
  sampleRef,
  studyIDSyn,
  verbose
)

Arguments

gdsProfile

an object of class gds.class (a GDS file), an opened Profile GDS file.

listPCA

a list containing the PCA object generated with the 1KG reference profiles (excluding the ones used to generate the synthetic data set) in an entry called "pca.unrel".

sampleRef

a vector of character strings representing the identifiers of the 1KG reference profiles that should not be used to create the reference PCA.

studyIDSyn

a character string corresponding to the study identifier. The study identifier must be present in the Profile GDS file.

verbose

a logical indicating if messages should be printed to show how the different steps in the function.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo GDS file is located in this package
dataDir <- system.file("extdata/demoKNNSynthetic", package="RAIDS")
fileProfileGDS <- file.path(dataDir, "ex1.gds")

## Open GDS files
gdsProfile <- openfn.gds(fileProfileGDS)

pca <- readRDS(file.path(dataDir, "pca1KG.RDS"))

## The function returns 0L when all parameters are valid
RAIDS:::validateComputePCAMultiSynthetic(gdsProfile=gdsProfile,
    listPCA=pca, sampleRef=c("HG00246", "HG00325"),
    studyIDSyn="MyStudy", verbose=FALSE)
#> [1] 0

## Close GDS file (it is important to always close the GDS files)
closefn.gds(gdsProfile)