Validate input parameters for computePCARefSample() function

This function validates the input parameters for the computePCARefSample function.

validateComputePCARefSample(
  gdsProfile,
  currentProfile,
  studyIDRef,
  np,
  algorithm,
  eigenCount,
  missingRate,
  verbose
)

Arguments

gdsProfile

an object of class gds.class, an opened Profile GDS file.

currentProfile

a single character string representing the profile identifier.

studyIDRef

a single character string representing the study identifier.

np

a single positive integer representing the number of CPU that will be used.

algorithm

a character string representing the algorithm used to calculate the PCA.

eigenCount

a single integer indicating the number of eigenvectors that will be in the output of the snpgdsPCA function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.

missingRate

a numeric value representing the threshold missing rate at with the SNVs are discarded; if NaN, no missing threshold.

verbose

a logical indicating if messages should be printed to show how the different steps in the function.

Value

The function returns 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(gdsfmt)

## Directory where demo GDS files are located
dataDir <- system.file("extdata", package="RAIDS")

## The GDS Sample (opened)
gdsSample <- openfn.gds(file.path(dataDir,
                    "GDS_Sample_with_study_demo.gds"), readonly=TRUE)

## The validation should be successful
RAIDS:::validateComputePCARefSample(gdsProfile=gdsSample,
    currentProfile="HCC01", studyIDRef="1KG", np=1L, algorithm="exact",
    eigenCount=32L, missingRate=0.02, verbose=FALSE)
#> [1] 0

## All GDS file must be closed
closefn.gds(gdsfile=gdsSample)