Validate the input parameters for computePoolSyntheticAncestry() function

The function validates the input parameters for the computePoolSyntheticAncestry function. When a parameter is not as expected, an error message is generated.

validateComputePoolSyntheticAncestry(
  gdsReference,
  profileGDS,
  profileID,
  dataRef,
  spRef,
  studyIDSyn,
  np,
  listCatPop,
  fieldPopIn1KG,
  fieldPopInfAnc,
  kList,
  pcaList,
  algorithm,
  eigenCount,
  missingRate
)

Arguments

gdsReference

an object of class gds.class (a GDS file), the opened 1KG GDS file.

profileGDS

an object of class gds.class (a GDS file), an opened Profile GDS file.

profileID

a single character string representing the profile identifier.

dataRef

a data.frame TODO

spRef

TODO

studyIDSyn

a character string corresponding to the study identifier. The study identifier must be present in the GDS Sample file.

np

a single positive integer representing the number of threads. Default: 1L.

listCatPop

a vector of character string representing the list of possible ancestry assignations. Default: ("EAS", "EUR", "AFR", "AMR", "SAS").

fieldPopIn1KG

a character string representing the TODO

fieldPopInfAnc

a character string representing the name of the column that will contain the inferred ancestry for the specified dataset. Default: "SuperPop".

kList

a vector of integer representing the list of values tested for the _K_ parameter. The _K_ parameter represents the number of neighbors used in the K-nearest neighbors analysis. If NULL, the value seq(2,15,1) is assigned. Default: seq(2,15,1).

pcaList

a vector of integer representing the list of values tested for the _D_ parameter. The _D_ parameter represents the number of dimensions used in the PCA analysis. If NULL, the value seq(2,15,1) is assigned. Default: seq(2,15,1).

algorithm

a character string representing the algorithm used to calculate the PCA. The 2 choices are "exact" (traditional exact calculation) and "randomized" (fast PCA with randomized algorithm introduced in Galinsky et al. 2016). Default: "exact".

eigen.cnt

a single integer indicating the number of eigenvectors that will be in the output of the snpgdsPCA function; if 'eigen.cnt' <= 0, then all eigenvectors are returned.

missing.rate

a numeric value representing the threshold missing rate at with the SNVs are discarded; the SNVs are retained in the snpgdsPCA with "<= missing.rate" only; if NaN, no missing threshold.

Value

The integer 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo 1KG GDS file is located in this package
dataDir <- system.file("extdata/tests", package="RAIDS")
fileGDS <- file.path(dataDir, "ex1_good_small_1KG_GDS.gds")
fileProfileGDS <- file.path(dataDir, "ex1_demo.gds")

## Open GDS files
gds1KG <- snpgdsOpen(fileGDS)
gdsProfile <- openfn.gds(fileProfileGDS)

dataRef <- data.frame(test=c(1,2), stringAsFactro=FALSE)

## The function returns 0L when all parameters are valid
RAIDS:::validateComputePoolSyntheticAncestry(gdsReference=gds1KG,
    profileGDS=gdsProfile, profileID="SampleID",
    dataRef=dataRef, spRef=NULL,  studyIDSyn="MyStudy",
    np=1L, listCatPop=c("EAS", "EUR", "AFR", "AMR", "SAS"),
    fieldPopIn1KG="SuperPop", fieldPopInfAnc="SuperPop",
    kList=seq(2,15,1), pcaList=seq(2,15,1),
    algorithm="exact", eigenCount=32L, missingRate=0.025)
#> [1] 0

## Close GDS files (it is important to always close the GDS files)
closefn.gds(gds1KG)
closefn.gds(gdsProfile)