Validate input parameters for computePoolSyntheticAncestryGr() function

This function validates the input parameters for the computePoolSyntheticAncestryGr function.

validateComputePoolSyntheticAncestryGr(
  gdsProfile,
  sampleRM,
  spRef,
  studyIDSyn,
  np,
  listCatPop,
  fieldPopInfAnc,
  kList,
  pcaList,
  algorithm,
  eigenCount,
  missingRate,
  verbose
)

Arguments

gdsProfile

an object of class SNPRelate::SNPGDSFileClass, the opened Profile GDS file.

sampleRM

a vector of character strings representing the identifiers of the population reference samples that should not be used to create the reference PCA.

spRef

a vector of character strings representing the known super population ancestry for the population reference profiles. The population reference profile identifiers are used as names for the vector.

studyIDSyn

a character string corresponding to the study identifier. The study identifier must be present in the GDS Sample file.

np

a single positive integer representing the number of threads.

listCatPop

a vector of character string representing the list of possible ancestry assignations.

fieldPopInfAnc

a character string representing the name of the column that will contain the inferred ancestry for the specified dataset.

kList

a vector of integer representing the list of values tested for the _K_ parameter. The _K_ parameter represents the number of neighbors used in the K-nearest neighbor analysis.

pcaList

a vector of integer representing the list of values tested for the _D_ parameter. The _D_ parameter represents the number of dimensions used in the PCA analysis.

algorithm

a character string representing the algorithm used to calculate the PCA. The 2 choices are "exact" (traditional exact calculation) and "randomized" (fast PCA with randomized algorithm introduced in Galinsky et al. 2016).

eigenCount

a single integer indicating the number of eigenvectors that will be in the output of the snpgdsPCA function; if 'eigenCount' <= 0, then all eigenvectors are returned.

missingRate

a numeric value representing the threshold missing rate at with the SNVs are discarded; the SNVs are retained in the snpgdsPCA only with "<= missingRate" only; if NaN, no missing threshold.

verbose

a logical indicating if message information should be printed.

Value

The function returns 0L when successful.

References

Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ, Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72. doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Directory where demo GDS files are located
dataDir <- system.file("extdata", package="RAIDS")

## The Profile GDS (opened)
gdsSample <- openfn.gds(file.path(dataDir,
                    "GDS_Sample_with_study_demo.gds"), readonly=TRUE)

## The known super population ancestry for the population reference profiles
spRef <- c("EUR", "SAS", "EAS", "EUR", "AFR")
names(spRef) <- c("HG00100", "HG00101", "HG00102", "HG00103", "HG00104")

## The validation should be successful
RAIDS:::validateComputePoolSyntheticAncestryGr(gdsProfile=gdsSample,
     sampleRM="TGCA_01", spRef=spRef,
     studyIDSyn="TCGA", np=1L, listCatPop=c("AFR", "EAS", "SAS"),
     fieldPopInfAnc="Pop", kList=seq_len(3),
     pcaList=seq_len(10), algorithm="exact", eigenCount=12L,
     missingRate=0.02, verbose=FALSE)
#> [1] 0

## All GDS file must be closed
closefn.gds(gdsfile=gdsSample)