Validate input parameters for createStudy2GDS1KG() function

This function validates the input parameters for the createStudy2GDS1KG function.

validateCreateStudy2GDS1KG(
  pedStudy,
  fileNameGDS,
  batch,
  studyDF,
  listSamples,
  PATHSAMPLEGDS,
  verbose
)

Arguments

pedStudy

a data.frame with those mandatory columns: "Name.ID", "Case.ID", "Sample.Type", "Diagnosis", "Source". All columns must be in character strings (no factor). The data.frame must contain the information for all the samples passed in the listSamples parameter. Only fileNamePED or pedStudy can be defined.

fileNameGDS

a character string representing the file name of the 1KG GDS file. The file must exist.

batch

a single positive integer representing the current identifier for the batch. Beware, this field is not stored anymore.

studyDF

a data.frame containing the information about the study associated to the analysed sample(s). The data.frame must have those 3 columns: "study.id", "study.desc", "study.platform". All columns must be in character strings (no factor).

listSamples

a vector of character string corresponding to the sample identifiers that will have a GDS Sample file created. The sample identifiers must be present in the "Name.ID" column of the RDS file passed to the fileNamePED parameter. If NULL, all samples in the fileNamePED are selected.

PATHSAMPLEGDS

a character string representing the path to the directory where the GDS Sample files will be created.

verbose

a logical indicating if message information should be printed.

Value

The function returns 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

gds1KG <- file.path(data.dir, "gds1KG.gds")

## The data.frame containing the information about the study
## The 3 mandatory columns: "study.id", "study.desc", "study.platform"
## The entries should be strings, not factors (stringsAsFactors=FALSE)
studyInfo <- data.frame(study.id="Pancreatic.WES",
                study.desc="Pancreatic study",
                study.platform="WES",
                stringsAsFactors=FALSE)

## PED Study
ped <- data.frame(Name.ID=c("Sample_01", "Sample_02"),
            Case.ID=c("TCGA-H01", "TCGA-H02"),
            Sample.Type=c("DNA", "DNA"),
            Diagnosis=c("Cancer", "Cancer"), Source=c("TCGA", "TCGA"))

## The validatiion should be successful
RAIDS:::validateCreateStudy2GDS1KG(pedStudy=ped, fileNameGDS=gds1KG,
            batch=1, studyDF=studyInfo,
            listSamples=c("Sample_01", "Sample_02"),
            PATHSAMPLEGDS=data.dir, verbose=TRUE)
#> [1] 0