Validate input parameters for getTableSNV() function

This function validates the input parameters for the getTableSNV function.

validateGetTableSNV(
  gdsRef,
  gdsSample,
  sampleCurrent,
  study.id,
  minCov,
  minProb,
  eProb,
  verbose
)

Arguments

gdsRef

an object of class gds.class (a GDS file), the opened 1KG GDS file (reference).

gdsSample

an object of class gds.class (a GDS file), the opened Sample GDS file.

sampleCurrent

a character string corresponding to the sample identifier used in pruningSample function.

study.id

a character string corresponding to the study identifier used in pruningSample function.

minCov

a single positive integer representing the minimum required coverage.

minProb

a single numeric between 0 and 1 representing the probability that the calculated genotype call is correct.

eProb

a single numeric between 0 and 1 representing the probability of sequencing error.

verbose

a logicial indicating if messages should be printed when the function is running.

Value

The integer 0L when the function is successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Directory where demo GDS files are located (refence and sample files)
data.dir <- system.file("extdata", package="RAIDS")

## The 1KG GDS file (opened) (reference file)
gds1KG <- openfn.gds(file.path(data.dir, "gds1KG.gds"), readonly=TRUE)

## The Sample GDS file (opened)
gdsSample <- openfn.gds(file.path(data.dir,
                         "GDS_Sample_with_study_demo.gds"), readonly=TRUE)

## The validation should be successful
RAIDS:::validateGetTableSNV(gdsRef=gds1KG, gdsSample=gdsSample,
     sampleCurrent="A101TCGA", study.id="TCGA", minCov=10L,
     minProb=0.998, eProb=0.002, verbose=TRUE)
#> [1] 0

## All GDS files must be closed
closefn.gds(gdsfile=gds1KG)
closefn.gds(gdsfile=gdsSample)