Validate input parameters for pruningSample() function

This function validates the input parameters for the pruningSample function.

validatePruningSample(
  gdsReference,
  method,
  currentProfile,
  studyID,
  listSNP,
  slideWindowMaxBP,
  thresholdLD,
  np,
  verbose,
  chr,
  superPopMinAF,
  keepPrunedGDS,
  pathProfileGDS,
  keepFile,
  pathPrunedGDS,
  outPrefix
)

Arguments

gdsReference

an object of class gds.class (a GDS file), the Population Reference GDS file.

method

a character string that represents the method that will be used to calculate the linkage disequilibrium in the snpgdsLDpruning() function.

currentProfile

a character string corresponding to the profile identifier used in LD pruning done by the snpgdsLDpruning() function. A Profile GDS file corresponding to the profile identifier must exist and be located in the pathProfileGDS directory.

studyID

a character string corresponding to the study identifier used in the snpgdsLDpruning function. The study identifier must be present in the Profile GDS file.

listSNP

a vector of SNVs identifiers specifying selected to be passed the the pruning function; if NULL, all SNVs are used in the snpgdsLDpruning function.

slideWindowMaxBP

a single positive integer that represents the maximum basepairs (bp) in the sliding window. This parameter is used for the LD pruning done in the snpgdsLDpruning function.

thresholdLD

a single numeric value that represents the LD threshold used in the snpgdsLDpruning function.

np

a single positive integer specifying the number of threads to be used.

verbose

a logicial indicating if information is shown during the process in the snpgdsLDpruning function.

chr

a character string representing the chromosome where the selected SNVs should belong. Only one chromosome can be handled. If NULL, the chromosome is not used as a filtering criterion.

superPopMinAF

a single positive numeric representing the minimum allelic frequency used to select the SNVs. If NULL, the allelic frequency is not used as a filtering criterion.

keepPrunedGDS

a logicial indicating if the information about the pruned SNVs should be added to the GDS Sample file.

pathProfileGDS

a character string representing the directory where the Profile GDS files will be created. The directory must exist.

keepFile

a logical indicating if RDS files containing the information about the pruned SNVs must be created.

pathPrunedGDS

a character string representing an existing directory. The directory must exist.

outPrefix

a character string that represents the prefix of the RDS files that will be generated. The RDS files are only generated when the parameter keepFile=TRUE.

Value

The function returns 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(gdsfmt)

## Directory where demo GDS files are located
dataDir <- system.file("extdata", package="RAIDS")

## The 1KG GDS file (opened)
gds1KG <- openfn.gds(file.path(dataDir,
            "PopulationReferenceDemo.gds"), readonly=TRUE)

## The validation should be successful
RAIDS:::validatePruningSample(gdsReference=gds1KG, method="corr",
     currentProfile="TGCA_01", studyID="TCGA",
     listSNP=c("sr10103", "sr10202"), slideWindowMaxBP=1000L,
     thresholdLD=0.008, np=1L, verbose=TRUE, chr=1,
     superPopMinAF=0.002, keepPrunedGDS=TRUE, pathProfileGDS=dataDir,
     keepFile=FALSE, pathPrunedGDS=dataDir, outPrefix="test")
#> [1] 0

## All GDS file must be closed
closefn.gds(gdsfile=gds1KG)