Validate input parameters for createAccuracyGraph function

This function validates the parameters for the createAccuracyGraph function.

validatecreateAUROCGraph(dfAUROC, title, selectD, selectColor)

Arguments

dfAUROC

a data.frame corresponding to res$paraSample$dfAUROC where res is the result of inferAncestry() or inferAncestryGeneAware() functions.

title

a character string representing the title of the graph.

selectD

a array of integer representing the selected PCA dimensions to plot. The length of the array cannot be more than 5 entries. The dimensions must tested by RAIDS (i.e. be present in the RDS file).

selectColor

a array of character strings representing the selected colors for the associated PCA dimensions to plot. The length of the array must correspond to the length of the selectD parameter. In addition, the length of the array cannot be more than 5 entries.

Value

The function returns 0L when successful.

Author

Astrid Deschênes and Pascal Belleau

Examples

## Path to RDS file with ancestry information generated by RAIDS (demo file)
dataDir <- system.file("extdata", package="RAIDS")
fileRDS <- file.path(dataDir, "TEST_01.infoCall.RDS")
info <- readRDS(fileRDS)
dfAUROC <- info$paraSample$dfAUROC

## Some of the column names must be updated to fit new standards
colnames(dfAUROC) <- c("D", "K", "Call", "L", "AUROC", "H")

## Validate parameters
RAIDS:::validatecreateAUROCGraph(dfAUROC=dfAUROC, title="Accuracy Graph", 
    selectD=c(6, 12), selectColor=c("blue","darkblue"))
#> [1] 0