RAIDS: Ancestry Inference from Cancer Sequencing

The RAIDS package implements specialized algorithms that enable ancestry inference from various cancer sequencing sources (RNA, Exome and Whole-Genome sequencing).

Details

The RAIDS package also implements simulation algorithm that generates synthetic cancer-derived data.

This code and analysis pipeline was designed and developed for the following publication:

References

Belleau, P, Deschênes A, Tuveson DA, Krasnitz A. Accurate and robust inference of genetic ancestry from cancer-derived molecular data across genomic platforms. bioRxiv 2022.02.01.478737 [link to bioRxiv](https://doi.org/10.1101/2022.02.01.478737)

See also

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Maintainer: Pascal Belleau <pascal_belleau@hotmail.com>