TODO

Create a data.frame with the allelic fraction TODO

computeAllelicFractionDNA(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  chrInfo,
  minCov = 10L,
  minProb = 0.999,
  eProb = 0.001,
  cutOffLOH = -5,
  cutOffHomoScore = -3,
  wAR = 9,
  verbose = FALSE
)

Arguments

gds

an object of class SNPRelate::SNPGDSFileClass, a SNP GDS file.

gdsSample

TODO

sampleCurrent

A character string corresponding to the sample identifier as used in pruningSample function.

study.id

A character string corresponding to the naome of the study as used in pruningSample function.

chrInfo

a vector chrInfo[i] = length(Hsapiens[[paste0("chr", i)]]) Hsapiens library(BSgenome.Hsapiens.UCSC.hg38)

minCov

a single positive integer representing the minimum required coverage. Default: 10L.

minProb

a single numeric between 0 and 1 representing TODO. Default: 0.999.

eProb

a single numeric between 0 and 1 representing the probability of sequencing error. Default: 0.001.

cutOffLOH

log of the score to be LOH . Default: -5.

cutOffHomoScore

TODO. Default: -3.

wAR

a single positive integer representing the size-1 of the window used to compute an empty box. Default: 9.

verbose

a logicial indicating if the function should print message when running. Default: FALSE.

Value

a data.frame with lap for the snv with depth > minCov. TODO

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO