Run a PCA analysis and a K nearest neighbor analysis on a small set of synthetic data

The function runs a PCA analysis using 1 synthetic profile from each sub-continental population. The reference samples used to create those synthetic profiles are first removed from the 1KG list of samples that generates the reference PCA. Then, the retained synthetic profiles are projected on the 1KG PCA space. Finally, a K nearest neighbor analysis using a range of K and D values is done.

computePoolSyntheticAncestryGr(
  gdsSample,
  sampleRM,
  spRef,
  study.id.syn,
  np = 1L,
  listCatPop = c("EAS", "EUR", "AFR", "AMR", "SAS"),
  fieldPopIn1KG = "superPop",
  fieldPopInfAnc = "SuperPop",
  kList = seq(2, 15, 1),
  pcaList = seq(2, 15, 1),
  algorithm = c("exact", "randomized"),
  eigen.cnt = 32L,
  missing.rate = 0.025
)

Arguments

gdsSample

an object of class SNPRelate::SNPGDSFileClass, the GDS Sample file.

sampleRM

a vector of character strings representing the identifiers of the 1KG reference samples that should not be used to create the reference PCA.

spRef

TODO

study.id.syn

a character string corresponding to the study identifier. The study identifier must be present in the GDS Sample file.

np

a single positive integer representing the number of threads. Default: 1L.

listCatPop

a vector of character string representing the list of possible ancestry assignations. Default: ("EAS", "EUR", "AFR", "AMR", "SAS").

fieldPopIn1KG

TODO. Default: "superPop".

fieldPopInfAnc

a character string representing the name of the column that will contain the inferred ancestry for the specified dataset. Default: "SuperPop".

kList

a vector of integer representing the list of values tested for the _K_ parameter. The _K_ parameter represents the number of neighbors used in the K-nearest neighbor analysis. If NULL, the value seq(2,15,1) is assigned. Default: seq(2,15,1).

pcaList

a vector of integer representing the list of values tested for the _D_ parameter. The _D_ parameter represents the number of dimensions used in the PCA analysis. If NULL, the value seq(2,15,1) is assigned. Default: seq(2,15,1).

algorithm

a character string representing the algorithm used to calculate the PCA. The 2 choices are "exact" (traditional exact calculation) and "randomized" (fast PCA with randomized algorithm introduced in Galinsky et al. 2016). Default: "exact".

eigen.cnt

a single integer indicating the number of eigenvectors that will be in the output of the snpgdsPCA function; if 'eigen.cnt' <= 0, then all eigenvectors are returned. Default: 32L.

missing.rate

a numeric value representing the threshold missing rate at with the SNVs are discarded; the SNVs are retained in the snpgdsPCA function with "<= missing.rate" only; if NaN, no missing threshold. Default: 0.025.

Value

a list TODO with the sample.id and eigenvectors and a table with KNN call for different K and pca dimension.

References

Galinsky KJ, Bhatia G, Loh PR, Georgiev S, Mukherjee S, Patterson NJ, Price AL. Fast Principal-Component Analysis Reveals Convergent Evolution of ADH1B in Europe and East Asia. Am J Hum Genet. 2016 Mar 3;98(3):456-72. doi: 10.1016/j.ajhg.2015.12.022. Epub 2016 Feb 25.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

# TODO
listEigenvector <- "TOTO"