Append the genotype information from a profile into the associated Profile GDS File

This function append the genotype information from a specific profile into the Profile GDS file. The genotype information is extracted from a SNV file as generated by SNP-pileup or other tools.

generateGDS1KGgenotypeFromSNPPileup(
  pathGeno,
  listSamples,
  listPos,
  offset,
  minCov = 10,
  minProb = 0.999,
  seqError = 0.001,
  dfPedProfile,
  batch,
  studyDF,
  pathProfileGDS,
  genoSource,
  verbose
)

Arguments

pathGeno

a character string representing the path to a directory with the genotype files for the profiles, as generated by snp-pileup. The path must exist.

listSamples

a vector of character string corresponding to the sample identifiers that will have a Profile GDS file created. The sample identifiers must be present in the "Name.ID" column of the data.frame passed to the dfPedProfile parameter.

listPos

a data.frame containing 2 columns. The first column, called "snp.chromosome" contains the name of the chromosome where the SNV is located. The second column, called "snp.position" contains the position of the SNV on the chromosome.

offset

a integer to adjust if the genome start at 0 or 1.

minCov

a single positive integer representing the minimum coverage needed to keep the SNVs in the analysis. Default: 10.

minProb

a single positive numeric between 0 and 1 representing the probability that the base change at the SNV position is not an error. Default: 0.999.

seqError

a single positive numeric between 0 and 1 representing the sequencing error rate. Default: 0.001.

dfPedProfile

a data.frame with the information about the sample(s). Those are mandatory columns: "Name.ID", "Case.ID", "Sample.Type", "Diagnosis", "Source". All columns must be in character strings. The data.frame must contain the information for all the samples passed in the listSamples parameter.

batch

a single positive integer representing the current identifier for the batch. Beware, this field is not stored anymore.

studyDF

a data.frame containing the information about the study associated to the analysed sample(s). The data.frame must have those 3 columns: "study.id", "study.desc", "study.platform". All columns must be in character strings.

pathProfileGDS

a character string representing the path to the directory where the GDS Sample files will be created.

genoSource

a character string with two possible values: 'snp-pileup' or 'generic'. It specifies if the genotype files are generated by snp-pileup (Facets) or are a generic format CSV file with at least those columns: 'Chromosome', 'Position', 'Ref', 'Alt', 'Count', 'File1R' and 'File1A'. The 'Count' is the depth at the specified position; 'FileR' is the depth of the reference allele and 'File1A' is the depth of the specific alternative allele.

verbose

a logical indicating if the function must print messages when running.

Value

The function returns 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the files in this package
dataDir <- system.file("extdata/tests", package="RAIDS")

## The data.frame containing the information about the study
## The 3 mandatory columns: "study.id", "study.desc", "study.platform"
## The entries should be strings, not factors (stringsAsFactors=FALSE)
studyDF <- data.frame(study.id = "MYDATA",
                        study.desc = "Description",
                        study.platform = "PLATFORM",
                        stringsAsFactors = FALSE)

## The data.frame containing the information about the samples
## The entries should be strings, not factors (stringsAsFactors=FALSE)
samplePED <- data.frame(Name.ID=c("ex1", "ex2"),
                    Case.ID=c("Patient_h11", "Patient_h12"),
                    Diagnosis=rep("Cancer", 2),
                    Sample.Type=rep("Primary Tumor", 2),
                    Source=rep("Databank B", 2), stringsAsFactors=FALSE)
rownames(samplePED) <- samplePED$Name.ID

## List of SNV positions
listPositions <- data.frame(snp.chromosome=c(rep(1, 10)),
    snp.position=c(3467333, 3467428, 3469375, 3469387, 3469502, 3469527,
    3469737, 3471497, 3471565, 3471618))

## Append genotype information to the Profile GDS file
result <- RAIDS:::generateGDS1KGgenotypeFromSNPPileup(pathGeno=dataDir,
            listSamples=c("ex1"), listPos=listPositions,
            offset=-1, minCov=10, minProb=0.999, seqError=0.001,
            dfPedProfile=samplePED, batch=1, studyDF=studyDF,
            pathProfileGDS=dataDir, genoSource="snp-pileup",
            verbose=FALSE)

## The function returns OL when successful
result
#> [1] 0

## The Profile GDS file 'ex1.gds' has been created in the
## specified directory
list.files(dataDir)
#>  [1] "1KG_Test.gds"                           
#>  [2] "1KG_Test_02.gds"                        
#>  [3] "Sample_Info_Test.RDS"                   
#>  [4] "ex1.gds"                                
#>  [5] "ex1.txt.gz"                             
#>  [6] "ex1_demo.gds"                           
#>  [7] "ex1_demo_with_pruning.gds"              
#>  [8] "ex1_demo_with_pruning_and_1KG_annot.gds"
#>  [9] "ex1_good_small_1KG_Annot_GDS.gds"       
#> [10] "ex1_good_small_1KG_GDS.gds"             

## Unlink Profile GDS file (created for demo purpose)
unlink(file.path(dataDir, "ex1.gds"))