Extract the genotype information for a SNV dataset using the GDS Sample file and the 1KG GDS file

The function generates a data.frame containing the genotype information from a initial list of SNVs associated to a specific sample. The function uses the information present in the 1KG GDS file and the GDS Sample file.

getTableSNV(
  gds,
  gdsSample,
  sampleCurrent,
  study.id,
  minCov = 10,
  minProb = 0.999,
  eProb = 0.001,
  verbose = FALSE
)

Arguments

gds

an object of class gds.class (a GDS file), the opened 1KG GDS file.

gdsSample

an object of class gds.class (a GDS file), the opened GDS Sample file.

sampleCurrent

a character string corresponding to the sample identifier used in pruningSample function.

study.id

a character string corresponding to the study identifier used in pruningSample function.

minCov

a single positive integer representing the minimum required coverage. Default: 10L.

minProb

a single numeric between 0 and 1 representing the probability that the calculated genotype call is correct. Default: 0.999.

eProb

a single numeric between 0 and 1 representing the probability of sequencing error. Default: 0.001.

verbose

a logicial indicating if messages should be printed when the function is running. Default: FALSE.

Value

a data.frame containing:

cnt.tot a single integer representing the total coverage for the SNV.

cnt.ref a single integer representing the coverage for the reference allele.

cnt.alt a single integer representing the coverage for the alternative allele.

snp.pos a single integer representing the SNV position.

snp.chr a single integer representing the SNV chromosome.

normal.geno a single numeric indicating the genotype of the SNV. The possibles are: 0 (wild-type homozygote), 1 (heterozygote), 2 (altenative homozygote), 3 indicating that the normal genotype is unknown.

snp.index a vector of integer representing the position of the SNVs in the 1KG GDS file.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Path to the demo pedigree file is located in this package
data.dir <- system.file("extdata", package="RAIDS")

## TODO