Extract the pruned SNVs in a population reference data set (ex:1KG) by chromosome and/or allelic frequency

The function extracts the pruned SNVs in a population reference data set (ex: 1KG) by chromosome and/or allelic frequency. The pruning is done through the linkage disequilibrium analysis. The pruned SNVs are saved in a RDS file.

pruning1KGbyChr(
  gdsReference,
  method = "corr",
  listSamples = NULL,
  slideWindowMaxBP = 5e+05,
  thresholdLD = sqrt(0.1),
  np = 1L,
  verbose = FALSE,
  chr = NULL,
  minAF = NULL,
  outPrefix = "pruned_1KG",
  keepObj = FALSE
)

Arguments

gdsReference

an object of class SNPRelate::SNPGDSFileClass, an opened SNP GDS file.

method

a character string that represents the method that will be used to calculate the linkage disequilibrium in the snpgdsLDpruning() function. The 4 possible values are: "corr", "r", "dprime" and "composite". Default: "corr".

listSamples

a character string that represents the method that will be used to calculate the linkage disequilibrium in the snpgdsLDpruning() function. The 4 possible values are: "corr", "r", "dprime" and "composite". Default: "corr".

slideWindowMaxBP

a single positive integer that represents the maximum basepairs (bp) in the sliding window. This parameter is used for the LD pruning done in the runLDPruning function. Default: 5e5.

thresholdLD

a single numeric value that represents the LD threshold used in the runLDPruning function. Default: sqrt(0.1).

np

a single positive integer specifying the number of threads to be used. Default: 1L.

verbose

a logical specifying if the function must provide more information about the process. Default: FALSE.

chr

a character string representing the chromosome where the selected SNVs should belong. Only one chromosome can be handled. If NULL, the chromosome is not used as a filtering criterion. Default: NULL.

minAF

a single positive numeric representing the minimum allelic frequency used to select the SNVs. If NULL, the allelic frequency is not used as a filtering criterion. Default: NULL.

outPrefix

a character string that represents the prefix of the RDS file(s) that will be generated. Default: "pruned_1KG".

keepObj

a logical specifying if the function must save the the processed information into a second RDS file. Default: FALSE.

Value

The function returns 0L when successful.

Author

Pascal Belleau, Astrid Deschênes and Alexander Krasnitz

Examples

## Required library
library(SNPRelate)

## Path to the demo pedigree file is located in this package
dataDir <- system.file("extdata", package="RAIDS")

## The 1KG Population Reference GDS demo file (opened)
gds1KG <- snpgdsOpen(file.path(dataDir, "1KG_Demo.gds"))

## The prefix of the RDS file to be created and containing the pruned SNVs
outPrefix <- "Pruned_Demo_Reference"

## Run only if directory in writing mode
if (file.access(getwd()) == 0 &&
    !file.exists(file.path(getwd(), paste0(outPrefix, ".rds")))) {

    ## Create a RDS file with the pruned SNVs
    RAIDS:::pruning1KGbyChr(gdsReference=gds1KG, outPrefix=outPrefix)

    prunedSNVs <- readRDS(file.path(getwd(), paste0(outPrefix, ".rds")))
    prunedSNVs
}
#> [1] "s2" "s7"

## Close 1K GDS file
closefn.gds(gds1KG)